# bplinvoice/openapi/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Invoicebackend API",
        "description": "Invoicebackend API documentation",
        "termsOfService": "",
        "contact": {
            "name": "",
            "url": "",
            "email": ""
        },
        "license": {
            "name": "unlicensed",
            "url": ""
        },
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
        "/api/way-bills/{id}": {
            "get": {
                "tags": [
                    "way-bill-resource"
                ],
                "operationId": "getWayBill",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/WayBillDTO"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "way-bill-resource"
                ],
                "operationId": "updateWayBill",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WayBillDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/WayBillDTO"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "way-bill-resource"
                ],
                "operationId": "deleteWayBill",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "patch": {
                "tags": [
                    "way-bill-resource"
                ],
                "operationId": "partialUpdateWayBill",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WayBillDTO"
                            }
                        },
                        "application/merge-patch+json": {
                            "schema": {
                                "$ref": "#/components/schemas/WayBillDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/WayBillDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CurrencyTypeDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    }
                }
            },
            "WayBillDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "boxLimit": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "shipmentType": {
                        "type": "string"
                    },
                    "opfac": {
                        "type": "string"
                    },
                    "deliveryAgent": {
                        "type": "string"
                    },
                    "estimatedReadyDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "currencyUom": {
                        "type": "string"
                    },
                    "estimatedShipDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "status": {
                        "type": "string"
                    },
                    "referenceNumber": {
                        "type": "string"
                    },
                    "currencyType": {
                        "$ref": "#/components/schemas/CurrencyTypeDTO"
                    }
                }
            },
...
        }
    }
}